+++ src/wetterdienst/api.py
@@ -31,6 +31,7 @@ class Wetterdienst:
             "dmo": "wetterdienst.provider.dwd.dmo.DwdDmoRequest",
             "road": "wetterdienst.provider.dwd.road.DwdRoadRequest",
             "swsmos": "wetterdienst.provider.dwd.swsmos.DwdSwsmosRequest",
+            "poi": "wetterdienst.provider.dwd.poi.DwdPoiRequest",
             "phenology": "wetterdienst.provider.dwd.phenology.DwdPhenologyRequest",
             "radar": "wetterdienst.provider.dwd.radar.DwdRadarValues",
             "alerts": "wetterdienst.provider.dwd.alerts.DwdWeatherAlertRequest",